permissions-center: move files to `enterprise/cmd/worker`.
Test plan:
CI.
Local sg run and check that permission syncs work.
