+++ enterprise/cmd/frontend/worker/auth/BUILD.bazel
@@ -2,44 +2,29 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "auth",
-    srcs = [
-        "perms_syncer_cleaner.go",
-        "perms_syncer_scheduler.go",
-        "sourcegraph_operator_cleaner.go",
-    ],
+    srcs = ["sourcegraph_operator_cleaner.go"],
     importpath = "github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/worker/auth",
     visibility = ["//visibility:public"],
     deps = [
         "//cmd/worker/job",
         "//cmd/worker/shared/init/db",
         "//enterprise/cmd/frontend/internal/auth/sourcegraphoperator",
-        "//enterprise/internal/authz",
         "//enterprise/internal/cloud",
-        "//enterprise/internal/database",
         "//internal/actor",
-        "//internal/api",
         "//internal/auth",
-        "//internal/conf",
         "//internal/database",
         "//internal/env",
         "//internal/errcode",
         "//internal/goroutine",
-        "//internal/metrics",
         "//internal/observation",
-        "//internal/timeutil",
         "//lib/errors",
         "@com_github_keegancsmith_sqlf//:sqlf",
-        "@com_github_sourcegraph_log//:log",
     ],
 )
 
 go_test(
     name = "auth_test",
-    srcs = [
-        "perms_syncer_cleaner_test.go",
-        "perms_syncer_scheduler_test.go",
-        "sourcegraph_operator_cleaner_test.go",
-    ],
+    srcs = ["sourcegraph_operator_cleaner_test.go"],
     embed = [":auth"],
     tags = [
         # Test requires localhost database
@@ -48,18 +33,10 @@ go_test(
     deps = [
         "//enterprise/cmd/frontend/internal/auth/sourcegraphoperator",
         "//enterprise/internal/cloud",
-        "//enterprise/internal/database",
         "//internal/auth",
-        "//internal/authz",
-        "//internal/conf",
         "//internal/database",
         "//internal/database/dbtest",
         "//internal/extsvc",
-        "//internal/timeutil",
-        "//internal/types",
-        "//schema",
-        "@com_github_google_go_cmp//cmp",
-        "@com_github_sourcegraph_log//:log",
         "@com_github_sourcegraph_log//logtest",
         "@com_github_stretchr_testify//assert",
         "@com_github_stretchr_testify//require",
