+++ enterprise/cmd/worker/internal/permissions/BUILD.bazel
@@ -5,6 +5,8 @@ go_library(
     srcs = [
         "bitbucket_projects.go",
         "config.go",
+        "perms_syncer_cleaner.go",
+        "perms_syncer_scheduler.go",
     ],
     importpath = "github.com/sourcegraph/sourcegraph/enterprise/cmd/worker/internal/permissions",
     visibility = ["//enterprise/cmd/worker:__subpackages__"],
@@ -12,10 +14,13 @@ go_library(
         "//cmd/frontend/globals",
         "//cmd/worker/job",
         "//cmd/worker/shared/init/db",
+        "//enterprise/cmd/worker/shared/permissions",
+        "//enterprise/internal/authz",
         "//enterprise/internal/database",
         "//internal/actor",
         "//internal/api",
         "//internal/authz",
+        "//internal/conf",
         "//internal/database",
         "//internal/database/basestore",
         "//internal/env",
@@ -25,7 +30,9 @@ go_library(
         "//internal/goroutine",
         "//internal/httpcli",
         "//internal/jsonc",
+        "//internal/metrics",
         "//internal/observation",
+        "//internal/timeutil",
         "//internal/types",
         "//internal/workerutil",
         "//internal/workerutil/dbworker",
@@ -44,6 +51,8 @@ go_test(
     srcs = [
         "bitbucket_projects_test.go",
         "main_test.go",
+        "perms_syncer_cleaner_test.go",
+        "perms_syncer_scheduler_test.go",
     ],
     data = glob(["testdata/**"]),
     embed = [":permissions"],
@@ -52,6 +61,7 @@ go_test(
         "requires-network",
     ],
     deps = [
+        "//enterprise/cmd/worker/shared/permissions",
         "//enterprise/internal/database",
         "//internal/api",
         "//internal/authz",
@@ -63,8 +73,11 @@ go_test(
         "//internal/extsvc",
         "//internal/extsvc/bitbucketserver",
         "//internal/observation",
+        "//internal/timeutil",
         "//internal/types",
         "//schema",
+        "@com_github_google_go_cmp//cmp",
+        "@com_github_sourcegraph_log//:log",
         "@com_github_sourcegraph_log//logtest",
         "@com_github_stretchr_testify//require",
         "@org_golang_x_exp//slices",
