+++ enterprise/cmd/worker/internal/permissions/perms_syncer_cleaner.go
@@ -1,4 +1,4 @@
-package auth
+package permissions
 
 import (
 	"context"
