+++ enterprise/cmd/worker/internal/permissions/perms_syncer_cleaner_test.go
@@ -1,4 +1,4 @@
-package auth
+package permissions
 
 import (
 	"context"
